+++ spec/helpers/stale_file_cleaner_spec.rb
@@ -1,7 +1,6 @@
-require 'spec_helper'
+require "spec_helper"
 
 describe "Cleans Stale Files" do
-
   it "removes files if they go over the limit" do
     file_size = 1000
     Dir.mktmpdir do |dir|
@@ -12,7 +11,7 @@ describe "Cleans Stale Files" do
       expect(old_file.exist?).to be_truthy
       expect(new_file.exist?).to be_truthy
 
-      ::LanguagePack::Helpers::StaleFileCleaner.new(dir).clean_over(2*file_size - 50)
+      ::LanguagePack::Helpers::StaleFileCleaner.new(dir).clean_over(2 * file_size - 50)
 
       expect(old_file.exist?).to be_falsey
       expect(new_file.exist?).to be_truthy
@@ -28,9 +27,9 @@ describe "Cleans Stale Files" do
 
       expect(old_file.exist?).to be_truthy
       expect(new_file.exist?).to be_truthy
-      dir_size = File.stat(dir)
+      File.stat(dir)
 
-      ::LanguagePack::Helpers::StaleFileCleaner.new(dir).clean_over(2*file_size + 50)
+      ::LanguagePack::Helpers::StaleFileCleaner.new(dir).clean_over(2 * file_size + 50)
 
       expect(old_file.exist?).to be_truthy
       expect(new_file.exist?).to be_truthy